requirement.txt
requests
pytest
gitpython
python-dotenv